Cleaned all WSACleanup and startup
Based on Windows's documentation, the entire app need at least one WSAStarup call, other gonna juste increment a counter, for wich the app need the same number of WSACleanup.
https://learn.microsoft.com/fr-fr/windows/win32/api/winsock/nf-winsock-wsastartup
To simplify, and to increase a lot the visibility, the app will call one WSAStartup, and one WSACleanup at all. This is done on CoreNetwork function.